Morse Code Project - 12/19/2020
Rounded out the Morse Game Program and did some small changes to the Morse Code Converter Program (now it just returns to the main menu when it is done, it doesn't ask if you want to convert something else. Also made a small fix to an mistake in the eng-to-morse-dict dictionary file.