Morse Code Program - 12/20/2020
Updated the Main Program so that now it won't accept entries other than 1 2 and 3 for the main menu. Updated the Morse Code Converter so that it has better flow and timing and also returns to the main menu directly after translating. Finished the Morse Code Game. Still need to do a little debugging, but it is 90% there. Added more phrases to Phrases.txt