Morse Code Program - 12/26/2020
Found some more bugs and resolved them all. The program is as finished as I think it can be for my current knowledge. It could probably be made more efficient with a little more work, but the improvement would be minimal and meddling with it any further could cause more problems than it solves.